requirements.txt
about-time==4.2.1
alive-progress==3.0.1
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
DateTime==4.9
et-xmlfile==1.1.0
exceptiongroup==1.0.4
grapheme==0.6.0
h11==0.14.0
idna==3.4
Jinja2==3.1.4
joblib==1.4.2
llvmlite==0.42.0
loguru==0.5.3
MarkupSafe==2.1.5
nose==1.3.7
notify-py==0.3.39
numba==0.59.1
numpy==1.24.0
openpyxl==3.0.10
outcome==1.2.0
packaging==22.0
pandas==1.5.2
pillow==11.0.0
pillow-avif-plugin==1.4.6
pycparser==2.21
PyIF==0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
scikit-learn==1.5.0
scipy==1.13.1
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
threadpoolctl==3.5.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
webdriver-manager==4.0.0
win32-setctime==1.1.0
wsproto==1.2.0
zope.interface==5.5.2
